requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
gradio>=4.0.0
pyttsx3>=2.90